refactor(example): move GreetingServiceBuilder
